Add files via upload
There is a rotary encoder attached to the end of the table for the X direction.  The encoder is read by an Arduino Uno and sent to a stepper motor attached to a rotary table.  As the x-direction handle is turned, the rotary table is turned in a ratio that allows spirals to be cut in a workpiece.  This version is hard-coded to cut a 23º spiral, which is what I need for the project that needed this.  I'll make a more generic version in the future.